requirements.txt
# Core
pytz==2018.4
Django==2.0.4
django-configurations==2.0
gunicorn==19.7.1
newrelic==3.2.0.91

# For the persistence stores
psycopg2-binary==2.7.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.1
django_unique_upload==0.2.1
django-polymorphic==2.0.2
django-hashid-field==2.1.1
django-pgviews==0.5.3

# Rest apis
djangorestframework==3.8.2
Markdown==2.6.11
django-filter==1.1.0
drfapikey==0.0.3
djangorestframework-jwt==1.11.0
django-cors-headers==2.2.0
simplejson==3.14.0

# Realtime Server
redis>=2.10
asyncio==3.4.3
aioredis==1.1.0
websockets==4.0.1

# Data Tools
numpy==1.14.2
pandas==0.22.0

# Developer Tools
ipdb==0.11
ipython==6.3.1
mkdocs==0.17.3
flake8==3.5.0
pudb==2017.1.4

# Testing
mock==2.0.0
factory-boy==2.10.0
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1

# Static and Media Storage
django-storages==1.6.6
boto3==1.7.7
